More methods use the class index, and two more bugs were fixed at client side: When an object was created in a collapsed node and what happened when a view has objects already deleted

+++ client/trunk/Communications/src/org/inventory/communications/core/LocalObjectLightImpl.java
@@ -105,7 +105,7 @@ public class LocalObjectLightImpl implements LocalObjectLight{ //This class impl
        if(obj == null)
            return false;
         if (obj.getClass().equals(this.getClass()))
-            return (this.getOid() == ((LocalObjectLightImpl)obj).getOid());
+            return (this.getOid().longValue() == ((LocalObjectLightImpl)obj).getOid().longValue());
         else
             return false;
    }
